Vendor pwsafer
Bug fixes, and API improvements will be necessary. For instance, the
Error is not exported in any capacity. And data is decrypted in memory
everything at once without any zeroize protection. And it does not
enforce verifying the MAC before access to data. And its mechanism for
data termination is a little dubious, trusts the last block to end the
file. And it puts decrypted data into a `Vec<u8>` for output, offering
no in-place (callback/Cursor) API.